+++ packages/create-sitecore-jss/src/templates/angular-sxp/src/app/routing/scripts/scripts.module.ts
@@ -1,0 +1,9 @@
+import { NgModule } from '@angular/core';
+import { ScriptsComponent } from './scripts.component';
+import { VisitorIdentificationComponent } from './visitor-identification/visitor-identification.component';
+
+@NgModule({
+  exports: [ScriptsComponent],
+  declarations: [ScriptsComponent, VisitorIdentificationComponent],
+})
+export class ScriptsModule {}